Add the QuickLight.js file
This is the initial (published) version of the QuickLight script.

When the yellow status marker is added to a token, it is given a light
source with a 40 unit light source, which dims at 20 units, and is
visible to all players. This will overwrite 'self-only' light, such as
that used to represent vision in complete darkness. Tokens which use
that type of vision should not use QuickLight.

Removal of the status marker turns off the light, by clearing the light
radius and dim light radius, and setting the light invisible to other
players.