Use a separate token for the light
Use a separate token, on the dynamic lighting layer ("walls"), so that
use of the quicklight does not interfere with innate token lighting.

Change the light on/off code to create/destroy the token instead.

When the owner token moves, move the light. Being on the lighting layer,
the torch image will not be seen (unless the GM is looking at the
lighting layer), and it will not animate when it is moved to follow its
owner token (as it would if it were on the token layer). This is what we
want. It does take a second for the lighting to update once the token
stops moving, but that may be due to the lighting system.