Add automatic token sizing
When a token is created or a character is changed, automatically set the
size of the token to match the size of the character, according to the
character sheet.

This does not work for swarms, so those will need to be manually
adjusted once they have been placed, until a solution is found to this.